moved RecruitmentApplicationForGangView to recruitment_views
