+++ backend/samfundet/view/recruitment_views.py
@@ -2,22 +2,26 @@ from __future__ import annotations
 
 from typing import Any
 
+from guardian.shortcuts import get_objects_for_user
+
+from rest_framework import status
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.decorators import action
-from rest_framework.permissions import DjangoModelPermissionsOrAnonReadOnly
+from rest_framework.permissions import IsAuthenticated, DjangoModelPermissionsOrAnonReadOnly
 
 from django.db.models import QuerySet
+from django.shortcuts import get_object_or_404
 from django.utils.decorators import method_decorator
 from django.views.decorators.csrf import ensure_csrf_cookie
 
 from root.utils.permissions import SAMFUNDET_VIEW_RECRUITMENT
 from root.custom_classes.permission_classes import RoleProtectedObjectPermissions, filter_queryset_by_permissions
 
-from samfundet.serializers import RecruitmentSerializer, RecruitmentGangSerializer, RecruitmentForRecruiterSerializer
+from samfundet.serializers import RecruitmentSerializer, RecruitmentGangSerializer, RecruitmentForRecruiterSerializer, RecruitmentApplicationForGangSerializer
 from samfundet.models.general import Gang
-from samfundet.models.recruitment import Recruitment
+from samfundet.models.recruitment import Recruitment, RecruitmentApplication
 
 # =============================== #
 #        Public views             #
@@ -64,4 +68,53 @@ class RecruitmentForRecruiterView(ModelViewSet):
         stats = self.get_object()
 
         serializer = self.serializer_class(stats)
+        return Response(serializer.data)
+
+
+class RecruitmentApplicationForGangView(ModelViewSet):
+    permission_classes = [IsAuthenticated]
+    serializer_class = RecruitmentApplicationForGangSerializer
+    queryset = RecruitmentApplication.objects.all()
+
+    # TODO: User should only be able to edit the fields that are allowed
+
+    @action(detail=True, methods=['put'])
+    def application_comment(self, request: Request, **kwargs: Any) -> Response:
+        application_id = kwargs.get('pk')
+
+        if 'application_comment' not in request.data:
+            return Response({'error': 'application_comment field is required'}, status=status.HTTP_400_BAD_REQUEST)
+
+        # Use direct update to bypass model save() method and signals
+        comment_text = request.data['application_comment']
+        updated = RecruitmentApplication.objects.filter(id=application_id).update(application_comment=comment_text)
+
+        if updated:
+            # Return the updated comment in the response
+            return Response({'application_comment': comment_text}, status=status.HTTP_200_OK)
+        return Response({'error': 'Application not found'}, status=status.HTTP_404_NOT_FOUND)
+
+    def list(self, request: Request) -> Response:
+        """Returns a list of all the recruitments for the specified gang."""
+        gang_id = request.query_params.get('gang')
+        recruitment_id = request.query_params.get('recruitment')
+
+        if not gang_id:
+            return Response({'error': 'A gang parameter is required'}, status=status.HTTP_400_BAD_REQUEST)
+
+        if not recruitment_id:
+            return Response({'error': 'A recruitment parameter is required'}, status=status.HTTP_400_BAD_REQUEST)
+
+        gang = get_object_or_404(Gang, id=gang_id)
+        recruitment = get_object_or_404(Recruitment, id=recruitment_id)
+
+        applications = RecruitmentApplication.objects.filter(
+            recruitment_position__gang=gang,
+            recruitment=recruitment,  # only include applications related to the specified recruitment
+        )
+
+        # check permissions for each application
+        applications = get_objects_for_user(user=request.user, perms=['view_recruitmentapplication'], klass=applications)
+
+        serializer = self.get_serializer(applications, many=True)
         return Response(serializer.data)
